removed unused GrantsShow component, renamed GrantsFinalizeShow to GrantsShow
