+++ src/Components/design/typography.css
@@ -27,7 +27,13 @@ h2,
 h3,
 h4,
 h5,
-h6 {
+h6,
+.heading-1,
+.heading-2,
+.heading-3,
+.heading-4,
+.heading-5,
+.heading-6 {
   margin: 3rem 0 1.38rem;
   font-weight: var(--font-weight-normal);
   line-height: 1.3;
@@ -38,32 +44,39 @@ h2,
 h5,
 h6,
 b,
-button {
+button,
+.bold {
   font-weight: var(--font-weight-bold);
 }
 
-h1 {
+h1,
+.heading-1 {
   margin-top: 0;
   font-size: var(--font-size-large-5);
 }
 
-h2 {
+h2,
+.heading-2 {
   font-size: var(--font-size-large-4);
 }
 
-h3 {
+h3,
+.heading-3 {
   font-size: var(--font-size-large-3);
 }
 
-h4 {
+h4,
+.heading-4 {
   font-size: var(--font-size-large-2);
 }
 
-h5 {
+h5,
+.heading-5 {
   font-size: var(--font-size-large-1);
 }
 
-h6 {
+h6,
+.heading-6 {
   text-transform: uppercase;
 }
